wip - current status is:
when we run:

otool -R -T -l examples/hello-world/zig-platform/dynhost | less

...we see that command 14 is LC_DATA_IN_CODE, which is correct.

however, command 15 is supposed to be LC_LOAD_DYLIB according to
otool, whereas we're actually seeing it as an unrecognized value:

- - - Unrecognized Mach-O command during linker preprocessing: 0x6162696c

Note that all the other commands are really low numbers (e.g.
LC_DATA_IN_CODE was 0x29), so this 0x6162696c seems like memory
garbage.

This suggests an offset problem or something, but we're not seeing
where the problem is. Maybe we should look for the 0xc that represents
LC_LOAD_DYLIB in the bytes following command 14, and see how much we're off.